@@ -1,9 +1,11 @@
 package com.android.updatesInMedicine;
 
 import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.cardview.widget.CardView;
+
 import android.os.Bundle;
-import android.support.v7.widget.CardView;
+//import androidx.appcompat.widget.CardView;
 import android.view.View;
 import android.widget.GridLayout;
 
