@@ -1,7 +1,8 @@
 package com.android.updatesInMedicine;
 
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import androidx.appcompat.app.AppCompatActivity;
+
 import com.android.androidrssfeedtutorial.R;
 
 public class AboutActivity extends AppCompatActivity {
@@ -12,5 +13,6 @@ public class AboutActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_about);
+        setTitle("About");
     }
 }
