@@ -1,20 +1,20 @@
 package com.android.updatesInMedicine;
 
 import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.widget.CardView;
 import android.view.View;
-import android.widget.Button;
 import android.widget.GridLayout;
-import android.widget.Toast;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.cardview.widget.CardView;
+
 
 import com.android.androidrssfeedtutorial.R;
 
 import java.util.ArrayList;
 
 
-public class JournalActivity extends AppCompatActivity {
+public class JournalActivity extends AppCompatActivity implements View.OnClickListener {
 
     ArrayList<String> rssLinks = new ArrayList<>();
     ArrayList<String> type = new ArrayList<>();
@@ -42,6 +42,19 @@ public class JournalActivity extends AppCompatActivity {
 
 
     }
+
+    @Override
+    public void onClick(View view) {
+        switch (view.getId()) {
+            case R.id.button:
+                startActivity(new Intent(JournalActivity.this, SubjectsList.class));
+                break;
+
+
+        }
+    }
+
+
 
     // we are setting onClickListener for each element
     private void setSingleEvent(GridLayout gridLayout) {
